Create utils module
