+++ platform-tools-sdk/cargo-build-sbf/src/post_processing.rs
@@ -1,5 +1,5 @@
 use {
-    crate::{rust_target_triple, spawn, Config},
+    crate::{spawn, utils::rust_target_triple, Config},
     log::{debug, error, info, warn},
     regex::Regex,
     solana_keypair::{write_keypair_file, Keypair},
